lib/igt_draw: Add 64bpp support for the non-XY_FAST_COLOR_BLT path
XY_COLOR_BLT can't do 64bpp since we have nowhere to store
the other half of the color. In order to support 64bpp
pixel formats we have to get a little bit creative and
use XY_PAT_BLT_IMMEDIATE instead. We simply treat the
buffer as 32bpp, double the horizontal dimensions to
compensate, and alternate between the lower and upper
dwords of the color value when filling the 8x8 pattern.

Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>