tests/kms_joiner: Reset the force_joiner_supported variable
With the current logic the test will fail on configs where the first
display supports force joiner and the second display doesn't supoprt it.
Reset the force_joiner_supported variable at the start of the loop to
avoid this failure.

Also remove the redundant check in all the force joiner subtests
for force_joiner_supported as the following output check is inherently
checking this as well.

Signed-off-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>