tests/core_setmaster: simplify device handling
Refactored test setup to dynamically determine the correct drm card
for testing. Instead of changing permissions for all cards,
change only one that will be tested.

Problem Description:
The test fails after running the `core_hotunplug@hot*` subtest, where
Card0 is populated as card1, This leads to a failure in the subsequent
`master-drop-set-user` test.

Command sequence for reproducing the issue:
- Check available cards: `ls /dev/dri/`
        - Output: by-path  card0  renderD128
- Run the test: `# ./core_hotunplug --r hotrebind-lateclose`
- Check again: ‘ls /dev/dri/’
        - Output after core_hotunplug : by-path  card1  renderD129
- Run the test: `# ./core_setmaster --r master-drop-set-user`
        - Output: gta@core_setmaster@master-drop-set-user failure

This failures on both XE and i915 for above sequence.

v2: corrected the e-mail address

Signed-off-by: Bommu Krishnaiah <krishnaiah.bommu@intel.com>
Cc: Emil Velikov <emil.l.velikov@gmail.com>
Cc: Himal Prasad Ghimiray <himal.prasad.ghimiray@intel.com>
Cc: Kamil Konieczny <kamil.konieczny@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>