tests/vmwgfx: Add test that draws triangle without using 3d commands.
The vmwgfx driver can be used in VMs with 3d disabled, in which case 3d
commands are not available, but buffers can still be created and used.
Validate buffers on such vms with a test that uses cairo to draw a
triangle on a vmwgfx mob and validate it with the same image in system
memory.

Signed-off-by: Maaz Mombasawala <maaz.mombasawala@broadcom.com>
Reviewed-by: Martin Krastev <martin.krastev@broadcom.com>