tests/intel/xe_module_load: Handling of snd_hda_intel
The problem: Running the following test-list twice fails on the second run:
igt@xe_module_load@load
igt@xe_module_load@unload

The load test will modprobe snd_hda_intel, while the unload will not rmmod
snd_hda_intel. Then running the load test the second time will fail as
snd_hda_intel is loaded and marked as an "unwanted_drivers".

It used to be the case that snd_hda_intel could prevent module load and unload
and we needed to take care of it in older kernels. This is no longer the case
so remove snd_had_intel from unwanted_drivers for this test.

Cc: Lucas De Marchi <lucas.demarchi@intel.com>
Signed-off-by: Peter Senna Tschudin <peter.senna@linux.intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>