tests/intel/xe_query: Clarify delta engine_cycles calculation
Let the compiler do the wrap around width instead of doing different
calculations depending on if the second value is lower than the first.

The only reason for ts2.engine_cycles to be lower than ts1.engine_cycles
is if a wrap around ocurred. The wrap around should happen
wrt the width reported by the kernel. However that is not really true
and kernel simply reports the value "as is". With Xe2, the value is
actually 64-bits and width is currently reported as 36 (that's being
changed in the kernel side). So, if engine_cycles would have to wrap
around 64 bits for that condition to hold true, and in that case we'd
calculate the delta by shifting (1 << ts2.width), giving a wrong result.

Luckily, wrapping 64 bits at 19.2Mhz takes forever and is not a problem
in real life.

Anyway, fix the math which also removes the ugly check by b >= a.

Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Link: https://lore.kernel.org/r/20241011030507.321961-3-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>