lib/intel_blt: Handle multiple ctrl-surf-copy blits on big surface
Copying ccs data for bigger surfaces requires couple of blits. This is
caused by the limitation of 10-bit size (for Xe and Xe2+) of the
ctrl-surf-copy operation. To handle bigger than 64MiB surfaces on Xe
and 4MiB on Xe2+ we need to divide blits to fit in mentioned 10-bit
size.

Change introduces looping around big surface to produce batch with
couple of ctrl-surf-copy blits.

Cc: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Link: https://lore.kernel.org/r/20241108064712.252927-2-zbigniew.kempczynski@intel.com
Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>