intel-ci: Blocklist igt@i915_module_load@load
This test expects the i915 module not loaded, and it SKIPs if it finds one
already loaded.  It was intended only for CI BAT runs as the very first
test executed.  It's not suitable for CI runs with random test selection
and ordering.

Since omitting it from non-BAT runs doesn't affect coverage, blocklist it
instead of maintaining as expected SKIP.

Closes: https://gitlab.freedesktop.org/drm/i915/kernel/-/issues/6227
Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>