tests/gem_mmap_offset: Fix OOM hits
The 'clear' subtest exercises correctness of object memory clearing on
passing a batch with the object to GPU for processing.  The exercise is
executed in several parallel threads, one per CPU.  Each thread repeats
the exercise in a time only limited loop, with no delay between
consecutive iterations.  In case of system memory objects, that happens
to exhaust all available physical memory, which is neither the goal nor
requirement of the exercise.

Make sure sufficient amount of physical memory is available before calling
another execbuf.

v2: Limit the scope of the fix to SMEM exercise.

Closes: https://gitlab.freedesktop.org/drm/i915/kernel/-/issues/11738
Link: https://lore.kernel.org/igt-dev/20241128111822.836089-6-janusz.krzysztofik@linux.intel.com/
Reviewed-by: Andi Shyti <andi.shyti@linux.intel.com>
Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>