lib/intel_compute: Use valid number of array size for unbind
Due to copy-paste number of elements for unbind is wrong for xe2.
Fix this as it should be paired with bind.

This is likely subject for refactor and use ARRAY_SIZE() but at the
moment it's a quickfix. Luckily number or entries value was smaller
than for bind so we didn't notice the failure, only there were some
dangling vmas in the vm.

Reviewed-by: Dominik Grzegorzek <dominik.grzegorzek@intel.com>
Link: https://lore.kernel.org/r/20241122102931.210761-1-zbigniew.kempczynski@intel.com
Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>