tests/intel/xe_eudebug_online: Add read/write pagefault online tests
Add read and write pagefault tests to xe_eudebug_online that checks if a
pagefault event is submitted by the KMD debugger when a pagefault occurs.

Test that read (load instruction) and write(store instruction) attempt to
load or store access to unallocated memory, causing a pagefault.
Examine the address causing the page fault and the number of eu threads
causing the pagefault.

v2: Integration to handle pagefault scenario in
    eu_attention_resume_trigger() callback. (Christoph)

Co-developed-by: Jonathan Cavitt <jonathan.cavitt@intel.com>
Signed-off-by: Gwan-gyeong Mun <gwan-gyeong.mun@intel.com>
Reviewed-by: Andrzej Hajda <andrzej.hajda@intel.com>
Link: https://lore.kernel.org/r/20241122143446.659440-5-gwan-gyeong.mun@intel.com