tests/intel/xe_fault_injection: Pick a valid PAT index for vm_bind
The vm_bind helper function in the fault injection test failed to
specify a PAT index to use for the bind, effectively leaving the index
as 0.  On Xe2 and Xe3 platforms, index 0 provides non-coherent behavior
which is incompatible with vm_bind operations on userptr, resulting in a
test failure before the test even got to the point of intentionally
injecting errors.

Fixes: 13e3d5126 ("tests/intel/xe_fault_injection: Inject errors during vm bind IOCTL")
Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/3499
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>