tests/intel/xe_wedged: Manage verbosity of GT reset fault injection
igt@xe_wedged@basic-wedged triggers a GT reset and uses fault injection
to ensure that the reset artificially fails.  This works as expected,
but depending on the value in the fail_gt_reset/verbose debugfs
when the test starts, different levels of output may be printed to
dmesg.  The test already marks the general "declared wedged" message as
expected so that igt_runner won't try to flag those as test errors, but
if verbose=2, then the fault injection framework will also issue print
dump_stack() output which gets flagged as a problem by igt_runner.

Explicitly set verbose=1 when running the test so that we get the
general wedged message and fault injection notification (which we
already tell igt_runner to ignore), but skip the unnecessary stack dump
that just confuses the runner.

Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/2919
Link: https://lore.kernel.org/r/20241126230240.1780940-1-matthew.d.roper@intel.com
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>