tests/vkms: Delete VKMS test list
First, this list is severely outdated. Therefore, it doesn't make sense
to keep an outdated test list in the repository, especially considering
that it is not being maintained.

Second, DRM CI now supports VKMS, and developers are maintaining an
updated list of IGT tests that VKMS supports. Thus, keeping this list
in IGT is redundant.

Signed-off-by: Maíra Canal <mcanal@igalia.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>