test/kms_content_protection: Account for reauthetication in LIC
During LIC the link may fail due to a variety of reasons but this
does not reflect a problem with HDCP. It becomes a problem only when
we are not able to recover from the Link Integrity Failure.
Add a check to see if CP property is CP_DESIRED and if so give it time
to reauthenticate and fail only if this reauthentication does not
happen.

--v2
-Add debug statement [Santhosh]
-Add closes tag [Santhosh]

Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/3508
Signed-off-by: Suraj Kandpal <suraj.kandpal@intel.com>
Reviewed-by: Santhosh Reddy Guddati <santhosh.reddy.guddati@intel.com>