tests/chamelium/kms_chamelium_edid: Limit the number of EDIDs tested to prevent CI timeout.
This patch reduces the number of non-4K HDMI EDIDs to 25 when
the extended flag is not used, as CI does not use this flag.
DP EDID tests are not reduced due to existing limited EDIDs.
The change ensures that only HDMI non-4K EDIDs are limited,
improving test efficiency and avoiding CI timeouts.

Cc: Mark Yacoub <markyacoub@chromium.org>
Cc: Mark Yacoub <markyacoub@google.com>
Cc: Louis Chauvet <louis.chauvet@bootlin.com>

Signed-off-by: Pranay Samala <pranay.samala@intel.com>
Reviewed-by: Kunal Joshi <kunal1.joshi@intel.com>