tests/kms_plane_scaling: get mode only after setting the output pipe
We find the mode before adding the dynamic intel-max-source-size
subtests because we don't want to add substests that don't have a
valid mode.  However, during the test itself, the mode info instance
is not valid anymore, because it is deallocated when setting the
output pipe.

To solve this, we need to get the mode info _after_ setting the output
pipe.  Avoid passing the mode to intel_max_source_size_test(), pass
the entire invalid_paramtests struct and let the function find the
correct mode info again.

Closes: https://gitlab.freedesktop.org/drm/i915/kernel/-/issues/8292
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-by: Swati Sharma <swati2.sharma@intel.com>