lib/amdgpu: Handle -ENODATA in amdgpu_wait_memory
The amdgpu_wait_memory function currently asserts if the return value
is non-zero and not -ECANCELED. However, -ENODATA is also a valid
error code that can be returned during GPU job timeout recovery,
particularly for queue resets. This patch updates the function to
also accept -ENODATA as a non-fatal error condition.

This change aligns with recent updates in the AMDGPU kernel driver
where -ENODATA is used to indicate queue-specific resets during
timeout recovery, while -ECANCELED or -ETIME is used for full GPU
resets. For more details, see the kernel discussion:
https://lists.freedesktop.org/archives/amd-gfx/2025-January/118795.html

Cc: Vitaly Prosyak <vitaly.prosyak@amd.com>
Cc: Christian Koenig <christian.koenig@amd.com>
Cc: Alexander Deucher <alexander.deucher@amd.com>

Signed-off-by: Jesse Zhang <jesse.zhang@amd.com>
Reviewed-by: Vitaly Prosyak <vitaly.prosyak@amd.com>