tests/intel/xe_exec_queue_property: Update first invalid property value
A new property will be added to support PXP [1], so we need to update
the invalid property testing to start from the next ID to avoid it from
failing once the PXP code is merged.

No new subtests are required to test the new property as that will be
covered by the PXP tests.

[1] https://lore.kernel.org/intel-xe/20250116001110.4158032-9-daniele.ceraolospurio@intel.com/
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Reviewed-by: Alan Previn <alan.previn.teres.alexis@intel.com>