tests/intel/kms_joiner: Add a new test to validate max-non-joiner mode
We need to ensure that the system does not enable a joiner for modes
that do not require it. This test runs on all pipes and validates
that the highest available resolution and refresh rate within the
max dot clock limit is selected. The test then forces a modeset and
flip on the given pipe and checks the status using the joiner check.
If the joiner is mistakenly enabled for a non-joiner mode, the test
should fail.

v2: Fix nonjoiner_mode_found to find the required case.
    Remove clk sort and minor fixes.
v3: Rename nonjoiner to non_joiner and minor modifications.
v4: Add joiner check.
v5: Rename test name and function name, Update test description,
    update non_joiner_mode_found logic.
v6: Update commit message and rename test, function name.
v7: Fix mode selection, loop through all pipes instead on
    pipe - 1, update igt description and add igt info to display
    the selected mode.

Signed-off-by: Jeevan B <jeevan.b@intel.com>
Reviewed-by: Karthik B S <karthik.b.s@intel.com>