tests/intel-ci: Add kms_sysfs_edid_timing to blocklist
Add kms_sysfs_edid_timing to blocklist. Test was written with the
idea that edid read shouldn't take an unreasonable amount of time.
The timeout is an arbitrary value because it depends on when the
display responds and, in some cases, especially when the display
is coming back from sleep, it can take quite a while. This test
has sporadic behavior across various platforms connected to
different displays.

Signed-off-by: Swati Sharma <swati2.sharma@intel.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Reviewed-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>