tests/intel/kms_cdclk: Add conditions to filter valid outputs
Add vrefresh as an additional parameter to check if highest and
lowest refresh rates are identical.
Skip the test, if highres and lowres are the same, as no cdclk
transition will occur.
Store highres and lowres of valid outputs in an array for reuse later.

v2:-fixed description (Ankit)
   -highres lowres stored in previous loop (Ankit)
   -removed comments (Ankit)
v3:-use i iterator (Ankit)
   -use func() for skip

Signed-off-by: Swati Sharma <swati2.sharma@intel.com>
Reviewed-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>