lib/intel_compute: Allow the user to provide a vm
By default the vm used in the compute library is created with default
flags and not visible externally. There can be situations where the
vm to be used already exists and requires special handling prior to
be used by the compute library. Add the possibility for the user to
provide such a vm when running a compute kernel. The place holder
structure containing the vm can later be extended to include more
parameters.

Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Signed-off-by: Francois Dugast <francois.dugast@intel.com>