tests/amdgpu/amd_replay: Skip test if no DRM DP AUX Interface
[Why]
The Replay IGT test need to communicate with eDP via drm_dp_aux*
Currently the test will report FAIL if no DRM DP AUX Interface
found on the system.

[How]
Skip the test if no DRM DP AUX Interface on the system and info
the tester the DRM DP AUX Interface need to be enabled in kernel
config to run the test.

The following setting in kernel config needs to be enabled for
amd_replay test.

Linux Kernel Configuration
└─>Device Drivers
└─>Graphics support
└─>Direct Rendering Manager
└─>DRM DP AUX Interface

Signed-off-by: Tom Chung <chiahsuan.chung@amd.com>
Reviewed-by: Alex Hung <alex.hung@amd.com>