tests/intel/kms_draw_crc: Skip formats not supported by the plane
The test only checks igt_display_has_format_mod() before
deciding to test a specific format. But that doesn't
mean the actual plane supports that format. Also check
for that once we know which plane to use.

Should turn the fp16 failures into skips on IVB because it
doesn't have fp16 support on the primary plane.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>