tests/intel/xe_oa: Rewrite the polling small buf test
The polling-small-buf test is specifically testing that POLLIN is still
set after an ENOSPC error. In the buggy code, POLLIN was cleared on
the ENOSPC error even though there was data to be read. The test was
putting an upper bound on the number of bytes read in a fixed duration
of time. While this worked, we had occassional failures that were due to
scheduling and other dependencies outside the scope of this test.

In an attempt to make the test more robust,
- use a more definitive method to wait for few reports
- force an ENOSPC error using a small buffer
- Ensure POLLIN is still set in the subsequent poll() call

v2:
- Add "why" in commit message (Ashutosh)
- Use uapi to wait for N reports

Signed-off-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>