tests/intel/xe_oa: Do not discard single reports in oa-formats
Internally the test is running a loop of 1000 iterations and calling
read in each iteration. The test is expecting a single read call to
return at least two reports. If two reports are not read, the logic
continues to the next iteration. This can easily fail if each read just
returns one report. In essence, many reports are available, but the
logic causes test to fail. Instead, count and use every report read
when waiting for two reports to be available.

Signed-off-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>