runner/kmemleak: Remove unnecessary lseek() calls
Earlier versions of the code performed multiple read operations on the
kmemleak file without closing and reopening the file descriptor. To
handle this, lseek() was used to reset the read position.

However, the current implementation no longer requires these lseek()
calls, as the redundant read operations have been removed. This commit
eliminates the leftover lseek() calls that are now unnecessary.

Cc: vitaly.prosyak@amd.com
Cc: christian.koenig@amd.com
Cc: alexander.deucher@amd.com
Cc: jesse.zhang@amd.com
Cc: harry.wentland@amd.com
Cc: zbigniew.kempczynski@intel.com
Cc: kamil.konieczny@linux.intel.com
Cc: ryszard.knop@intel.com
Cc: lucas.demarchi@intel.com
Cc: katarzyna.piecielska@intel.com
Signed-off-by: Peter Senna Tschudin <peter.senna@linux.intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>