tests/intel/i915_module_load: Do not skip on snd
The fact that snd_hda_* is loaded doesn't mean i915 is. It used to be
the case that loading e.g. snd_hda_intel would load i915 because of a
request_module("i915") in the kernel. However that changed in
commit e6d0c13e9f46 ("ALSA: hda: i915: Remove extra argument from
snd_hdac_i915_init").

Drop the snd_* modules from unwanted_drivers.

Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Closes: https://gitlab.freedesktop.org/drm/i915/kernel/-/issues/13896
Link: https://lore.kernel.org/r/20250311195339.1206647-1-lucas.demarchi@intel.com
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>