drm-uapi/amdgpu: align with kernel commit adc218676eef
Align with kernel commit adc218676eef
commit dc6a198ce587c0469136c54888f58cb2ec2582f1
Merge: fa6355bbe92e adc218676eef
Merge tag 'v6.12' into amd-staging-drm-next
Linux 6.12
https://lore.kernel.org/dri-devel/20240906211008.3072097-1-alexander.deucher@amd.com/T/#u?utm_source)
Commit adc218676eef: https://github.com/torvalds/linux/commit/adc218676eef

This patch introduces new UAPI/IOCTL for usermode graphics
queue. IGT test cases fill this structure and request
the graphics driver to add a graphics work queue for it.
The output of this UAPI is a queue id.

This UAPI maps the queue into GPU, so the graphics app can start
submitting work to the queue as soon as the call returns.

v2: Kamil requested the SHA of a commit from the drm-next branch.

Signed-off-by: Sunil Khatri <sunil.khatri@amd.com>
Reviewed-by: Vitaly Prosyak <vitaly.prosyak@amd.com>